Crate concurrency_manager

The concurrency manager is responsible for concurrency control of transactions.

The concurrency manager contains a lock table in memory. Lock information can be stored in it and reading requests can check if these locks block the read.

In order to mutate the lock of a key stored in the lock table, it needs to be locked first using lock_key or lock_keys.

Modules

key_handle

lock_table

Structs

ConcurrencyManager

KeyHandle
An entry in the in-memory table providing functions related to a specific key.

KeyHandleGuard
A KeyHandle with its mutex locked.

LockTable